Fix bugs related to Fn:Equals and Fn:If evaluation
The Fn:Equals now can compare different types of ValueNode objs
depending on their actual value attribute. Still needs a better way
to override the equals function in order for this to work in a
proper way. The Fn:If has been fixed to encode only the needed nodes
and to take carry over the type of the subproperty that the fun
needs to evaluate to.