Remove anything related to Aws Managed Policies
Removed the JSON with all the policies. Removed the usage of that JSON.
Introduced only a set of Aws Managed Policies ARNs to be used only to detected
if an ARN points to a Managed Policy. We don't need the text of the policy at
this stage, but the ARN is only used to give a name to a very generic External
Resource.